chaged long_int to size_t

+++ src/SWI-Prolog.h
@@ -882,7 +882,7 @@ PL_EXPORT(int)	PL_wchars_to_term(const pl_wchar_t *chars,
 
 PL_EXPORT(int)		PL_is_initialised(int *argc, char ***argv);
 PL_EXPORT(int)		PL_initialise(int argc, char **argv);
-PL_EXPORT(int)		PL_boot_from_memory(int argc, char **argv, char *mem_data, long int mem_size);
+PL_EXPORT(int)		PL_boot_from_memory(int argc, char **argv, char *mem_data, size_t mem_size);
 #ifdef __CYGWIN__
 PL_EXPORT(void)		PL_install_readline(void);
 #else
